refactor(frontend): extract game engine from vue component
